Refactor client folder
