+++ client/components/CoreComponents/allproducts.js
@@ -1,10 +1,10 @@
 import React, {useEffect, useState} from 'react'
 import {connect} from 'react-redux'
-import {fetchAllProducts, fetchSingleProduct} from '../store/products'
+import {fetchAllProducts, fetchSingleProduct} from '../../store/products'
 import ReactPaginate from 'react-paginate'
 import {Link} from 'react-router-dom'
 import FilterProducts from './FilterProducts'
-import {createNewOpenOrder} from '../store/orders'
+import {createNewOpenOrder} from '../../store/orders'
 
 const AllProducts = props => {
   const {products, addToCart, getProducts, getSingleProduct} = props
