+++ client/components/UserAccountComponents/OrderHistoryDetails.js
@@ -1,6 +1,6 @@
 import React, {Component} from 'react'
 import {connect} from 'react-redux'
-import {fetchMyCompletedOrder} from '../store/orders'
+import {fetchMyCompletedOrder} from '../../store/orders'
 import {Link} from 'react-router-dom'
 import moment from 'moment'
