+++ client/components/index.js
@@ -3,9 +3,9 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './navbar'
-export {default as Home} from './Home'
-export {default as LogIn} from './LogIn'
-export {default as SignUp} from './SignUp'
-export {default as MyUserAccount} from './MyUserAccount'
-export {default as EditMyAccount} from './EditMyAccount'
+export {default as Navbar} from './CoreComponents/navbar'
+export {default as Home} from './CoreComponents/Home'
+export {default as LogIn} from './UserAccountComponents/LogIn'
+export {default as SignUp} from './UserAccountComponents/SignUp'
+export {default as MyUserAccount} from './UserAccountComponents/MyUserAccount'
+export {default as EditMyAccount} from './UserAccountComponents/EditMyAccount'
