+++ client/routes.js
@@ -2,16 +2,17 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Link, Redirect, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {LogIn, SignUp, Home, MyUserAccount, EditMyAccount} from './components'
+import {LogIn, SignUp, MyUserAccount, EditMyAccount} from './components'
+import Home from './components/CoreComponents/Home'
 import {getMe} from './store'
-import AllProducts from './components/allproducts'
-import SingleProduct from './components/singleProduct'
-import MyCart from './components/MyCart'
+import AllProducts from './components/CoreComponents/allproducts'
+import SingleProduct from './components/CoreComponents/singleProduct'
+import MyCart from './components/CartComponents/MyCart'
 import AdminDashboard from './components/AdminComponents/AdminDashboard'
-import Confirmation from './components/confirmation'
-import OrderHistoryDetails from './components/OrderHistoryDetails'
-import GuestCheckoutForm from './components/GuestCheckoutForm'
-import PageNotFound from './components/PageNotFound'
+import Confirmation from './components/CoreComponents/confirmation'
+import OrderHistoryDetails from './components/UserAccountComponents/OrderHistoryDetails'
+import GuestCheckoutForm from './components/UserAccountComponents/GuestCheckoutForm'
+import PageNotFound from './components/CoreComponents/PageNotFound'
 
 /**
  * COMPONENT
